Add files via upload
cam creates a phantom webcam for services like Zoom. The script is executed at the command which requires file path specification then provides user input for three prompts to be displayed overlay to the image provided. It sizes the image to 1024x768 for  best compatibility. And the text is displayed over it. This is ideal for before and after service needs during Zoom attendance.